+++ data-raw/packages.R
@@ -47,7 +47,8 @@ paletteer_packages <- tibble::tribble(
   "werpals", "sciencificity/werpals", "0.1.0", FALSE, NA,
   "calecopal", "an-bui/calecopal", "0.1.0", FALSE, NA,
   "colorblindr", "clauswilke/colorblindr", "0.1.0", FALSE, NA,
-  "colRoz", "jacintak/colRoz", "0.2.2", FALSE, NA
+  "colRoz", "jacintak/colRoz", "0.2.2", FALSE, NA,
+  "unikn", "hneth/unikn", "0.2.0.9003", TRUE, "0.2.0"
 ) %>%
   arrange(Name)